+++ packages/ui-chart/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.19](https://github.com/nativescript-community/ui-chart/compare/v2.0.18...v2.0.19) (2024-05-23)
+
+**Note:** Version bump only for package @nativescript-community/ui-chart
+
+## [2.0.18](https://github.com/nativescript-community/ui-chart/compare/v2.0.17...v2.0.18) (2024-05-23)
+
+**Note:** Version bump only for package @nativescript-community/ui-chart
+
+## [2.0.17](https://github.com/nativescript-community/ui-chart/compare/v2.0.16...v2.0.17) (2024-05-21)
+
+**Note:** Version bump only for package @nativescript-community/ui-chart
+
+## [2.0.16](https://github.com/nativescript-community/ui-chart/compare/v2.0.15...v2.0.16) (2024-05-21)
+
+**Note:** Version bump only for package @nativescript-community/ui-chart
+
+## [2.0.15](https://github.com/nativescript-community/ui-chart/compare/v2.0.14...v2.0.15) (2024-05-21)
+
+**Note:** Version bump only for package @nativescript-community/ui-chart
+
+## [2.0.14](https://github.com/nativescript-community/ui-chart/compare/v2.0.13...v2.0.14) (2024-05-21)
+
+**Note:** Version bump only for package @nativescript-community/ui-chart
+
+## [2.0.13](https://github.com/nativescript-community/ui-chart/compare/v2.0.12...v2.0.13) (2024-05-03)
+
+**Note:** Version bump only for package @nativescript-community/ui-chart
+
 ## [2.0.12](https://github.com/nativescript-community/ui-chart/compare/v2.0.11...v2.0.12) (2024-04-30)
 
 **Note:** Version bump only for package @nativescript-community/ui-chart